backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
timm==0.9.12
numpy==1.24.3
python-dotenv==1.0.0
supabase==2.0.3
